service/requirements.txt
# Common
django==3.2.5
djangorestframework==3.11.2
psycopg2-binary==2.8.6
whitenoise==5.1.0
requests==2.24.0
Pillow==8.3.2
elasticsearch==7.8.1
python-dateutil==2.8.1
suds-community==0.8.5
uritemplate==3.0.1
Jinja2==2.11.3
django-x-robots-tag-middleware==1.3.3

# AWS
awscli==1.18.195
boto3==1.16.35
requests-aws4auth==1.0
django-storages==1.9.1

# Tools
ipython==7.16.1
tqdm==4.50.2
invoke==1.4.1
ipdb==0.13.3
sentry-sdk==1.0.0
beautifulsoup4==4.9.1

-r dependencies.txt
